refactor into canvas component
