+++ src/components/Chart.js
@@ -15,18 +15,6 @@ function createData(time, amount) {
   return { time, amount };
 }
 
-const data = [
-  createData("00:00", 12),
-  createData("03:00", 11),
-  createData("06:00", 12),
-  createData("09:00", 9),
-  createData("12:00", 13),
-  createData("15:00", 12),
-  createData("18:00", 11.5),
-  createData("21:00", 10.5),
-  createData("24:00", undefined),
-];
-
 export default function Chart() {
   const theme = useTheme();
